idaholab/Malcolm#530, work in progress on netbox plugin installation
