+++ netbox/supervisord.conf
@@ -39,7 +39,6 @@ command=/opt/netbox/venv/bin/python /usr/local/bin/netbox_init.py
   --token "%(ENV_SUPERUSER_API_TOKEN)s"
   --library "%(ENV_NETBOX_DEVICETYPE_LIBRARY_IMPORT_PATH)s"
   --preload "%(ENV_NETBOX_PRELOAD_PATH)s"
-  --custom-plugins "%(ENV_NETBOX_CUSTOM_PLUGINS_PATH)s"
   --postgres-host "%(ENV_DB_HOST)s"
   --postgres-db "%(ENV_DB_NAME)s"
   --postgres-user "%(ENV_DB_USER)s"
